requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
jinja2>=3.1.4
apscheduler>=3.10.4
feedparser>=6.0.11
python-dotenv>=1.0.1
